# docs/swagger/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cognitive Quorum API",
    "description": "Backend for Cognitive Quorum application.",
    "version": "2.1.0"
  },
  "paths": {
...
    "/workflows": {
      "post": {
        "tags": [
          "Orchestration"
        ],
        "summary": "Create Workflow",
        "description": "Creates a new workflow definition in the database.\n\nArgs:\n    request (ExecutionWorkflowCreateRequest): The workflow payload containing name and steps.\n    engine (WorkflowEngine): The workflow engine dependency.\n\nReturns:\n    dict: The status and generated workflow_id.",
        "operationId": "create_workflow_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionWorkflowCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A confirmation object with the new Workflow ID.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executions": {
      "post": {
        "tags": [
          "Orchestration"
        ],
        "summary": "Start Execution",
        "description": "Initiates a new workflow execution asynchronously.\n\nSupports Multipart/Form-Data for optional file uploads alongside JSON inputs.\n\nArgs:\n    request (Request): The raw FastAPI request (for parsing file keys).\n    json_payload (str): The raw JSON string containing execution config.\n    background_tasks (BackgroundTasks): Logic for handling async operations.\n    engine (WorkflowEngine): The workflow engine dependency.\n    current_user (CurrentUserDep): The authenticated user initiating the run.\n\nReturns:\n    dict: The status and execution_id.",
        "operationId": "execute_workflow_executions_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_execute_workflow_executions_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The ID of the newly started execution background job.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executions/recent": {
      "get": {
        "tags": [
          "Orchestration"
        ],
        "summary": "List Recent Executions",
        "description": "Retrieves a list of the most recent workflow executions (Scoped by Org).",
        "operationId": "get_recent_executions_executions_recent_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Maximum number of executions to return.",
              "default": 5,
              "title": "Limit"
            },
            "description": "Maximum number of executions to return."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by execution status (e.g., 'completed', 'failed').",
              "title": "Status"
            },
            "description": "Filter by execution status (e.g., 'completed', 'failed')."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of recent execution records, sorted by time (descending).",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/executions/{execution_id}": {
      "get": {
        "tags": [
          "Orchestration"
        ],
        "summary": "Get Execution Status",
        "description": "Retrieves the full status and result data for a specific execution ID.\n\nPerforms on-the-fly hydration of legacy result structures if necessary.",
        "operationId": "get_execution_status_executions__execution_id__get",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The UUID of the execution to retrieve.",
              "title": "Execution Id"
            },
            "description": "The UUID of the execution to retrieve."
          }
        ],
        "responses": {
          "200": {
            "description": "The detailed status, result, and state of a specific execution.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executions/{execution_id}/retry": {
      "post": {
        "tags": [
          "Orchestration"
        ],
        "summary": "Retry Execution",
        "description": "Resumes a failed, rejected, or interrupted execution from its last successful state.",
        "operationId": "retry_execution_executions__execution_id__retry_post",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The UUID of the execution to retry.",
              "title": "Execution Id"
            },
            "description": "The UUID of the execution to retry."
          }
        ],
        "responses": {
          "200": {
            "description": "Confirmation that the execution is resuming.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_execute_workflow_executions_post": {
        "properties": {
          "json_payload": {
            "type": "string",
            "title": "Json Payload"
          }
        },
        "type": "object",
        "required": [
          "json_payload"
        ],
        "title": "Body_execute_workflow_executions_post"
      },
...
      "ExecutionWorkflowCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The unique, human-readable name for the new workflow."
          },
          "steps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Steps",
            "description": "A sequential list of step configurations defining the workflow logic."
          }
        },
        "type": "object",
        "required": [
          "name",
          "steps"
        ],
        "title": "ExecutionWorkflowCreateRequest",
        "description": "Payload for creating a new workflow definition.\n\nAttributes:\n    name (str): Human-readable name.\n    steps (list[dict]): List of step configurations."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}